change some logics (add train and predict files)
