--- configs/run_conf.yaml
@@ -1,1 +1,0 @@
-my_run: train
