+++ models/metrics/RandomForestClassifier.json
@@ -1,1 +1,1 @@
-{"accuracy_score": 0.8, "f1_score": 0.8125000000000001, "r2_score": 0.19642857142857162, "mae": 0.2, "rmse": 0.4472135954999579}
+{"accuracy_score": 0.85, "f1_score": 0.8524590163934426, "r2_score": 0.3973214285714287, "mae": 0.15, "rmse": 0.3872983346207417}
